fix issues with changing use case
